refactor: move constant declarations into relevant functions and files for better organization
