--- internal/core/core_kube.go
@@ -1,68 +1,0 @@
-package core
-
-import (
-	"context"
-	"os"
-	"sync"
-
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
-)
-
-var kubeConfigMap sync.Map
-
-func kubeConfig(ctx context.Context, facility FacilityRepo, action ActionRepo, uuid, name string) (*rest.Config, error) {
-	key := uuid + "/" + name
-
-	if v, ok := kubeConfigMap.Load(key); ok {
-		return v.(*rest.Config), nil
-	}
-
-	config, err := newKubeConfig(ctx, facility, action, uuid, name)
-	if err != nil {
-		return nil, err
-	}
-
-	kubeConfigMap.Store(key, config)
-
-	return config, nil
-}
-
-func newKubeConfig(ctx context.Context, facility FacilityRepo, action ActionRepo, uuid, name string) (*rest.Config, error) {
-	// kubernetes-control-plane
-	leader, err := facility.GetLeader(ctx, uuid, name)
-	if err != nil {
-		return nil, err
-	}
-
-	result, err := runAction(ctx, action, uuid, leader, "get-kubeconfig", nil)
-	if err != nil {
-		return nil, err
-	}
-
-	configAPI, err := clientcmd.Load([]byte(result.Output["kubeconfig"].(string)))
-	if err != nil {
-		return nil, err
-	}
-
-	config, err := clientcmd.NewDefaultClientConfig(*configAPI, &clientcmd.ConfigOverrides{}).ClientConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	// Write CA data to temp file for helm
-	if config.CAData != nil {
-		tmpFile, err := os.CreateTemp("", "otterscale-ca-*.crt")
-		if err != nil {
-			return nil, err
-		}
-		defer tmpFile.Close()
-
-		if _, err := tmpFile.Write(config.CAData); err != nil {
-			return nil, err
-		}
-		config.CAFile = tmpFile.Name()
-	}
-
-	return config, nil
-}
